requirements.txt
asgiref==3.8.1
Django==5.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==25.1
scipy==1.16.1
setuptools==78.1.1
sqlparse==0.5.2
typing_extensions==4.12.2
wheel==0.45.1
whitenoise==6.9.0
